+++ CHANGES.rst
@@ -21,6 +21,9 @@ development version
 * Similarly, ``expressed_D.tab`` and ``expressed_J.tab`` and their
   ``.pdf`` counterparts are created in each iteration.
 * Removed ``parse`` subcommand (functionality is in the ``igblast`` subcommand)
+* New CDR3 detection method (only heavy chain sequences): CDR3 start/end coordinates
+  are pre-computed using the database V and J sequences. Increases detection rate
+  to 99% (previously less than 90%).
 
 v0.8.0 (2017-06-20)
 -------------------